requirements.txt
fastapi
uvicorn[standard]
groq
pydantic
pydantic-settings
qdrant-client
sentence-transformers
pandas
openpyxl
python-jose[cryptography]
python-multipart
langchain
langchain-core
langchain-groq
langchain-community
langchain-text-splitters
presidio-analyzer
presidio-anonymizer
httpx
requests